Add support for node's light selection during IPR. Fixes #441
Changing light selection on geo nodes during IPR was not supported.
This support is now added, although when changing the categories
attribute on a light node, is not yet supported. Following the same
way for lights, would be trickier and more expensive (assuming that
the nodes might be greater in number), so we might need to find a
better way for it.