Fixed render region loss during IPR issue. Closes #442.
When switching between cam tabs or modifying one of the active cameras
parameters, render region would be updated with the crop values set on
that camera rather than respecting the crop made from iDisplay. With
this fix the crop region made from i-display will not be lost, unless
changing one of the crop attributes' values from the camera.