Fixed multi-light ROP error. Closes #453.
On commit eb7451d3 we were dealing with VDB visibility on AOV when
vdb was being used as phantom. Also we added another check for this and
tried to show the AOV only for VDBs with emission grid. The second one
was causing errors since we were trying to get vdb grids for
non-existing primitives. This was as a result of wrong vdb filename or
when vdb was being modifed by another node. On the other hand, the
previous fix for phantom VDBs will work with no problems.