Prevented "VDB loaders" from being instanced at the SOP level.
Either the OBJ is a VDB loader, and we handle transforms ourselves, or it is
not, and we let Houdini refine it for us.  The previous code was instantiating
VDB loaders, which resulted in their transform being applied twice. This didn't
matter before we added support for transformed VDB loaders (in
46823014c142ceb00a19a41f1ada865d1a94ac41 ), because the transform that was
applied twice was the identity.

Should fix issue 456.