Fixed very stupid memory corruption.
"Shooting oneself in the foot, a case study."

This std::vector was introduced in 5d8bfcb3ec0553c41ccb74d79b20f79cdcc3500e with
the explicit goal of keeping std::strings local in order to avoid pointing to
strings located inside Houdini primitives. However, since we accumulated the
addresses of the strings while the vector was being filled, those were being
invalidated when the vector was resized & re-allocated, which completely
defeated the purpose of that strings buffer.

It does seem to be more of a problem on Windows, which is probably due to the
vector being re-allocated more frequently on that platform.

Using a std::deque should guarantee that its contents won't be moved
arbitrarily, which will keep the references intact.