Improved motion blur specification for Alembic procedural.
The procedural now also uses the shutter interval to decide which frames to pick
up from the Alembic file. Since it also continues to use the "fps" and "frame"
parameters (as a time multiplier and a time offset, respectively, for
compatibility with other plugins that export 0-centered frames) we set thsoe to
their neutral value of 1 and 0.