Revert to the old way of multi-light handling in 3Delight. Reverts 848b5981
We will be rolling back to the old way of handling multi-lights. Instead
of having one single checkbox, which would render all lights and groups
of lights in separate layers, we now have a list of all lights from
where users can select which one they want to be rendered as a separate
layer.