Stopped forcing the Cryptomatte preview layer to be 32-bit.
This layer is not used for looking up IDs : it's only a preview of the areas
with the same ID meant for the human eye. Plus, it's been deprecated in the
Cryptomatte specification. It will now use the bit depth selected by the user
(which could be 32), just like the other layers.