Added Visibility tab on dlAOVGroup node.
Also simplified add_and_connect_aov_group() function on vop.cpp where we
were checking all material's input to find if an aovGroup is connected.
We know that one input can only accept one node so instead we now just look
on aovGroup input by name and check if the node connected to it is the
dlAOVGroup node.