Added shutter offset attributes to the Houdini camera.
This allows the user to shift the opening and closing time of the shutter so
they both occur earlier or later. In order to keep the UI simple, we provide 3
presets, to which the actual (and hidden) numerical offset is linked. An
advanced user could decide to break this link (therefore ignoring the selected
preset) and control the numerical value directly.