Added support for Cryptomatte AOVs in Deep EXR output.
Otherwise, the IDs just get output to a regular one-channel layer which, apart
from hitting a small bug in the 3Delight library (the crash described in issue
https://gitlab.com/3Delight/3delight-core/-/issues/157), is of no use to Nuke's
Cryptomatte plugin.

So, here we are : however strange that might seem, we now output a
depth-simulating Cryptomatte ID into a non-deep layer of an otherwise deep EXR
file.